+++ apptest.go
@@ -12,6 +12,7 @@ import (
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 	corev1 "k8s.io/api/core/v1"
+	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -86,6 +87,7 @@ func New(config Config) (*AppSetup, error) {
 		// be able to watch for runtime objects of a certain type.
 		appSchemeBuilder := runtime.SchemeBuilder(schemeBuilder{
 			v1alpha1.AddToScheme,
+			apiextensionsv1.AddToScheme,
 		})
 		err = appSchemeBuilder.AddToScheme(scheme.Scheme)
 		if err != nil {
@@ -147,6 +149,19 @@ func (a *AppSetup) InstallApps(ctx context.Context, apps []App) error {
 	err = a.waitForDeployedApps(ctx, apps)
 	if err != nil {
 		return microerror.Mask(err)
+	}
+
+	return nil
+}
+
+// EnsureCRDs will register the passed CRDs in the k8s API used by the client.
+func (a *AppSetup) EnsureCRDs(ctx context.Context, crds []*apiextensionsv1.CustomResourceDefinition) error {
+	var err error
+	for _, crd := range crds {
+		err = a.ctrlClient.Create(ctx, crd)
+		if err != nil {
+			return microerror.Mask(err)
+		}
 	}
 
 	return nil