+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Expose method `EnsureCRDs` to register CRDs in the k8s API.
+
 ## [0.4.1] - 2020-10-30
 
 ### Added